adding modifications for making strucutre similar to dsb
