+++ egs/madcat_arabic/v1/local/check_dependencies.sh
@@ -1,0 +1,8 @@
+#!/usr/bin/env bash
+
+
+set -e  # exit on error
+
+scripts/dependencies/check_pytorch.sh
+
+scripts/dependencies/check_packages.sh
