+++ egs/madcat_arabic/v1/run.sh
@@ -1,39 +1,47 @@
-#!/usr/bin/env bash
+#!/bin/bash
 
-set -e
+set -e # exit on error
+. ./path.sh
+
 stage=0
+
+. parse_options.sh  # e.g. this parses the --stage option if supplied.
+
+
+. ./cmd.sh ## You'll want to change cmd.sh to something that will work on your system.
+           ## This relates to the queue.
+
 nj=70
 download_dir1=/export/corpora/LDC/LDC2012T15/data
 download_dir2=/export/corpora/LDC/LDC2013T09/data
 download_dir3=/export/corpora/LDC/LDC2013T15/data
 
-. ./cmd.sh ## You'll want to change cmd.sh to something that will work on your system.
-           ## This relates to the queue.
-. ./path.sh
-. parse_options.sh  # e.g. this parses the above options
-                      # if supplied.
+local/check_dependencies.sh
 
-./local/check_tools.sh
-
-mkdir -p data/{train,test,dev}/data
-mkdir -p data/local/{train,test,dev}
-mkdir -p data/{train,test,dev}/masks
-data_splits_dir=data/download/data_splits
 
 if [ $stage -le 0 ]; then
-  echo "$0: Downloading data splits..."
-  echo "Date: $(date)."
-  local/download_data.sh --data_splits $data_splits_dir
+  # data preparation
+  local/prepare_data.sh
 fi
 
-if [ $stage -le 0 ]; then
-  for dataset in test dev train; do
-    echo "$0: Extracting mask from page image for dataset:  $dataset. "
-    echo "Date: $(date)."
-    dataset_file=$data_splits_dir/madcat.$dataset.raw.lineid
-    local/extract_masks.sh --nj $nj --cmd $cmd --dataset_file $dataset_file \
-                           --download_dir1 $download_dir1 --download_dir2 $download_dir2 \
-                           --download_dir3 $download_dir3 data/local/$dataset
-  done
-fi
+
+#epochs=10
+#depth=5
+#dir=exp/unet_${depth}_${epochs}_sgd
+#if [ $stage -le 1 ]; then
+#  # training
+#  local/run_unet.sh --epochs $epochs --depth $depth
+#fi
+#
+#if [ $stage -le 2 ]; then
+#    echo "doing segmentation...."
+#  local/segment.py \
+#    --dir $dir \
+#    --train-dir data/train_val \
+#    --train-image-size 128 \
+#    --core-config $dir/configs/core.config \
+#    --unet-config $dir/configs/unet.config \
+#    $dir/model_best.pth.tar
+#
+#fi
